Fix resize/alloc in RegionHeader, improve API
It doesn't make that much sense to write a chunk to a file without the header
using the same file, especially considering that the function uses the header
to find the offset, which means that the file object must also have a 2kib
header! And if that's the case, shouldn't the RegionHeader just be on that file
then??? See, it doesn't make any sense! It's better now.

Resize/alloc were pretty bad; fixed some major bugs in there. It was a mess...